requirements.txt
streamlit>=1.44.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
google-generativeai>=0.8.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
requests>=2.31.0
fpdf2>=2.8.0
